Separated sample usage of the library inside main.c. Added docstrings to the C functions and added section in the ReadMe.
