+++ c/matching/recursive_match.c
@@ -1,10 +1,22 @@
-#include <stdio.h>
-#include <stdbool.h>
 #include "recursive_match.h"
 #include "matrix.h"
 
+/**
+ * @brief Creates and initializes a Matcher object for matrix matching.
+ * 
+ * Allocates memory for a Matcher structure and initializes its parameters, 
+ * including the matrix, axis, limit flag, and the matches array based on 
+ * the axis (0 for rows, 1 for columns). It also calculates the matrix's 
+ * minimum value.
+ * 
+ * @param matrix Pointer to the Matrix structure used for matching.
+ * @param axis Axis along which to perform matching (0 for rows, 1 for cols).
+ * @param limit Boolean flag to limit matching based on the minimum matrix value.
+ * 
+ * @return A pointer to the created Matcher structure, or NULL on error.
+ */
 Matcher* create_matcher(Matrix* matrix, int axis, bool limit) {
-    // Allocate memory for the Matcher structure
+    // Allocate memory for the Matcher structure.
     Matcher* matcher = (Matcher*)malloc(sizeof(Matcher));
     if (!matcher) {
         fprintf(stderr, "\t - [ERROR] Memory allocation failed for matcher.\n");
@@ -24,7 +36,7 @@ Matcher* create_matcher(Matrix* matrix, int axis, bool limit) {
             return NULL;
     }
 
-    // Dynamically allocate memory for matches
+    // Dynamically allocate memory for matches.
     int* matches = (int*)malloc(size * sizeof(int));
     if (!matches) {
         fprintf(stderr, "\t - [ERROR] Memory allocation failed for matches!\n");
@@ -46,11 +58,25 @@ Matcher* create_matcher(Matrix* matrix, int axis, bool limit) {
     return matcher;
 }
 
+/**
+ * @brief Performs rematching for a specific element in the matrix.
+ * 
+ * This function finds the maximum value in the given items (row/column of 
+ * the matrix), and attempts to match the current element (index i) with 
+ * the best match (index of maximum value). It also handles conflicts by 
+ * recursively rematching if necessary. The match may be limited by the 
+ * matrix's minimum value.
+ * 
+ * @param matcher Pointer to the Matcher structure containing matrix and other 
+ *                matching parameters.
+ * @param i Index of the current element to match.
+ * @param items Array of float values representing a row/column of the matrix.
+ */
 void rematch(Matcher* matcher, int i, float* items) {
     int max_index = 0;
     float max_value = items[0];
 
-    // Find the argmax
+    // Find the argmax.
     for (int k = 0; k < matcher->size; k++) {
         if (items[k] > max_value) {
             max_index = k;
@@ -61,7 +87,7 @@ void rematch(Matcher* matcher, int i, float* items) {
     if (matcher->limit && max_value <= matcher->min_value)
         return;
 
-    // Find the index containing max_index
+    // Find the index containing max_index.
     bool contains_max_index = false;
     int j = 0;
     for (int k = 0; k < matcher->size; k++) {
@@ -80,8 +106,10 @@ void rematch(Matcher* matcher, int i, float* items) {
             duplicate = matcher->matrix->data[max_index][j];
         
         if (duplicate < max_value) {
-            matcher->matches[j] = -1; // Unmatch previous because current match is a better fit.
-            matcher->matches[i] = max_index; // Match the current column/row with the row/column.
+            // Unmatch previous because current match is a better fit.
+            matcher->matches[j] = -1; 
+            // Match the current column/row with the row/column.
+            matcher->matches[i] = max_index; 
             
             float *new_items;
             if (matcher->axis == 0)
@@ -97,7 +125,7 @@ void rematch(Matcher* matcher, int i, float* items) {
         int min_index = 0;
         float min_value = items[0];
 
-        // Find the argmin
+        // Find the argmin.
         for (int k = 0; k < matcher->size; k++) {
             if (items[k] < min_value) {
                 min_index = k;
@@ -109,7 +137,7 @@ void rematch(Matcher* matcher, int i, float* items) {
         // the second highest value as the next potential match.
         items[max_index] = min_value - 1.0;
 
-        // If i has not changed to j, rematch current match;
+        // If i has not changed to j, rematch current match.
         rematch(matcher, i, items);
 
     } else {
@@ -118,6 +146,16 @@ void rematch(Matcher* matcher, int i, float* items) {
 
 }
 
+/**
+ * @brief Runs the matching process for all elements along the specified axis.
+ * 
+ * Iterates through all elements along the specified axis, calling the 
+ * rematch function for each. Frees the memory allocated for the row/column 
+ * of the matrix after each match operation.
+ * 
+ * @param matcher Pointer to the Matcher structure containing matrix and 
+ *                matching parameters.
+ */
 void run(Matcher* matcher) {
     for (int i = 0; i < matcher->size; i++) {
         float *items;
@@ -126,6 +164,6 @@ void run(Matcher* matcher) {
         else 
             items = get_column(matcher->matrix, i);
         rematch(matcher, i, items);
-        free(items); // Free allocated memory for the items
+        free(items); // Free allocated memory for the items.
     }
 }
